+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmObjectHelper.kt
@@ -19,9 +19,9 @@ package io.realm.internal
 import io.realm.BaseRealmObject
 import io.realm.EmbeddedObject
 import io.realm.MutableRealm
-import io.realm.RealmInstant
 import io.realm.RealmList
 import io.realm.RealmObject
+import io.realm.UpdatePolicy
 import io.realm.dynamic.DynamicMutableRealmObject
 import io.realm.dynamic.DynamicRealmObject
 import io.realm.internal.dynamic.DynamicUnmanagedRealmObject
@@ -38,6 +38,7 @@ import io.realm.internal.platform.realmObjectCompanionOrThrow
 import io.realm.internal.schema.ClassMetadata
 import io.realm.internal.schema.PropertyMetadata
 import io.realm.internal.schema.RealmStorageTypeImpl
+import io.realm.internal.schema.realmStorageType
 import io.realm.internal.util.Validation.sdkError
 import io.realm.schema.RealmStorageType
 import kotlin.reflect.KClass
@@ -212,7 +213,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
         value: BaseRealmObject?,
-        updatePolicy: MutableRealm.UpdatePolicy = MutableRealm.UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
         cache: ObjectCache = mutableMapOf()
     ) {
         obj.checkValid()
@@ -267,7 +268,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         col: String,
         list: RealmList<Any?>,
-        updatePolicy: MutableRealm.UpdatePolicy = MutableRealm.UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
         cache: ObjectCache = mutableMapOf()
     ) {
         val existingList = getList<T>(obj, col)
@@ -287,7 +288,7 @@ internal object RealmObjectHelper {
     internal fun assign(
         target: BaseRealmObject,
         source: BaseRealmObject,
-        updatePolicy: MutableRealm.UpdatePolicy,
+        updatePolicy: UpdatePolicy,
         cache: ObjectCache
     ) {
         if (target is DynamicRealmObject) {
@@ -306,7 +307,7 @@ internal object RealmObjectHelper {
     internal fun assignTyped(
         target: BaseRealmObject,
         source: BaseRealmObject,
-        updatePolicy: MutableRealm.UpdatePolicy,
+        updatePolicy: UpdatePolicy,
         cache: ObjectCache
     ) {
         val metadata: ClassMetadata = target.realmObjectReference!!.metadata
@@ -367,7 +368,7 @@ internal object RealmObjectHelper {
     internal fun assignDynamic(
         target: DynamicMutableRealmObject,
         source: BaseRealmObject,
-        updatePolicy: MutableRealm.UpdatePolicy,
+        updatePolicy: UpdatePolicy,
         cache: ObjectCache
     ) {
         val properties: List<Pair<String, Any?>> = if (source is DynamicRealmObject) {
@@ -472,7 +473,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectReference<out BaseRealmObject>,
         propertyName: String,
         value: R,
-        updatePolicy: MutableRealm.UpdatePolicy = MutableRealm.UpdatePolicy.ERROR,
+        updatePolicy: UpdatePolicy = UpdatePolicy.ERROR,
         cache: ObjectCache = mutableMapOf()
     ) {
         obj.checkValid()
@@ -611,14 +612,5 @@ internal object RealmObjectHelper {
             CollectionType.RLM_COLLECTION_TYPE_LIST -> "RealmList<$elementTypeString>"
             else -> TODO("Unsupported collection type: $collectionType")
         }
-    }
-
-    private fun <T : Any> KClass<T>.realmStorageType(): KClass<*> = when (this) {
-        RealmInstantImpl::class -> RealmInstant::class
-        DynamicRealmObject::class,
-        DynamicUnmanagedRealmObject::class,
-        DynamicMutableRealmObject::class ->
-            BaseRealmObject::class
-        else -> this
     }
 }