+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/dynamic/DynamicMutableRealmImpl.kt
@@ -18,7 +18,7 @@ package io.realm.internal.dynamic
 
 import io.realm.BaseRealmObject
 import io.realm.Deleteable
-import io.realm.MutableRealm
+import io.realm.UpdatePolicy
 import io.realm.dynamic.DynamicMutableRealm
 import io.realm.dynamic.DynamicMutableRealmObject
 import io.realm.internal.BaseRealmImpl
@@ -62,7 +62,7 @@ internal open class DynamicMutableRealmImpl(
     // differentiate the dynamic objects without bloating the type space
     override fun copyToRealm(
         obj: BaseRealmObject,
-        updatePolicy: MutableRealm.UpdatePolicy
+        updatePolicy: UpdatePolicy
     ): DynamicMutableRealmObject {
         return io.realm.internal.copyToRealm(configuration.mediator, realmReference, obj, updatePolicy, mutableMapOf()) as DynamicMutableRealmObject
     }